requirements.txt
absl-py==0.8.1
alphashape==1.0.2
asn1crypto==0.24.0
astor==0.8.1
autopep8==1.4.4
backcall==0.1.0
cachetools==4.0.0
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
Click==7.0
click-log==0.3.2
cmake==3.15.3
colorama==0.4.1
colorthief==0.2.1
cycler==0.10.0
Cython==0.29.14
dask==2.5.2
decorator==4.4.0
dlib==19.19.0
gast==0.2.2
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
hypopt==1.0.9
idna==2.6
imageio==2.6.1
imutils==0.5.3
ipython==7.9.0
ipython-doctest-plugin==0.1
ipython-genutils==0.2.0
jedi==0.15.1
joblib==0.13.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
llvmlite==0.30.0
lxml==4.5.0
Markdown==3.2.1
matplotlib==3.1.1
menpo==0.10.0
menpofit==0.6.1
missingno==0.4.2
mpldatacursor==0.6.2
networkx==2.3
nose==1.3.7
numba==0.46.0
numpy==1.17.1
oauthlib==3.1.0
opencv-contrib-python==3.4.2.16
opencv-python==3.4.2.16
opt-einsum==3.1.0
ordereddict==1.1
pandas==0.25.1
parso==0.5.1
patsy==0.5.1
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==6.1.0
procrustes==0.2.1
prompt-toolkit==2.0.10
protobuf==3.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.18
Pygments==2.4.2
PyJWT==1.5.3
pyparsing==2.4.2
python-dateutil==2.8.0
python-gflags==3.1.2
pytz==2019.2
PyWavelets==1.0.3
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.1
scikit-learn==0.21.3
scipy==1.4.1
seaborn==0.9.0
Shapely==1.6.4.post2
six==1.14.0
skan==0.8
sklearn==0.0
statsmodels==0.11.0
tensorboard==2.1.0
tensorflow-estimator==2.1.0
tensorflow-gpu==2.1.0
tensorflow-gpu-estimator==2.1.0
termcolor==1.1.0
toolz==0.10.0
tqdm==4.42.1
traitlets==4.3.3
urllib3==1.22
utils==0.9.0
wcwidth==0.1.7
Werkzeug==1.0.0
wrapt==1.12.0
xlrd==1.2.0
yattag==1.12.2
